requirements.txt
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0
snowflake-connector-python==3.12.2 
slack_bolt
